% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{addforecast}
\alias{addforecast}
\alias{addforecast.ccamforecast}
\alias{addforecast.forecastset}
\title{CCAM add forecasts}
\usage{
addforecast(fit, what, dotcol = "black", dotpch = 19, dotcex = 1.5,
  intervalcol = gray(0.5, alpha = 0.5), ...)

\method{addforecast}{ccamforecast}(fit, what, dotcol = "black", dotpch = 19,
  dotcex = 1.5, intervalcol = gray(0.5, alpha = 0.5), ...)

\method{addforecast}{forecastset}(fit, what, dotpch = 19, dotcex = 1.5,
  ci = TRUE, col = c("#332288", "#88CCEE", "#44AA99", "#117733", "#999933",
  "#DDCC77", "#661100", "#CC6677", "#882255", "#AA4499"), ...)
}
\arguments{
\item{fit}{the object returned from ccam.fit}

\item{what}{what to plot}

\item{dotcol}{color for dot}

\item{dotpch}{pch for dot}

\item{dotcex}{cex for dot}

\item{intervalcol}{color for interval}

\item{...}{extra arguments not currently used}
}
\description{
CCAM add forecasts
}
\details{
internal plotting fun
}
