% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp.R
\name{avail}
\alias{avail}
\title{check availability of certain objects}
\usage{
avail(x)
}
\arguments{
\item{x}{object (character)}
}
\description{
check availability of certain objects
}
\details{
x can be either MP (management procedure) or IE (implementation error)
}
