% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bubble}
\alias{bubble}
\alias{heat}
\title{Bubble plot}
\usage{
bubble(x, xlab = "Year", ylab = "Age", scale = 15, col = c("black",
  "darkgreen"), alpha = 0.8)

heat(x, high = c("grey", "darkgreen"), low = "darkred", xlab = "Year",
  ylab = "Age", posneg = FALSE)
}
\arguments{
\item{x}{matrix}

\item{xlab}{x label}

\item{ylab}{y label}

\item{scale}{scaler for the bubble size (scale_range)}

\item{col}{color for the bubbles. if 2 colors are given, they indicate positive and negative values}

\item{alpha}{transparency of the bubbles}

\item{x}{matrix}

\item{col}{colors that indicate the color scale limit}

\item{ncol}{number of colors to use in plot}

\item{leg}{logical (plot legend?)}

\item{leground}{number of digits to round legend}
}
\description{
Bubble plot

Heat plot
}
