% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{c.ccamforecast}
\alias{c.ccamforecast}
\title{Collect ccamforecast objects}
\usage{
\method{c}{ccamforecast}(...)
}
\arguments{
\item{...}{ccam forecasts to be combined}
}
\description{
Collect ccamforecast objects
}
\details{
...
}
