\name{canmackConf}
\alias{canmackConf}
\docType{data}
\title{
canmackConf
}
\description{
canmackConf
}
\usage{data("canmackConf")}
\format{
 The format is:
 $ minAge                   
 $ maxAge                   
 $ maxAgePlusGroup          
 $ keyLogFsta               
 $ corFlag                  
 $ keyLogFpar               
 $ keyQpow                  
 $ keyVarF                  
 $ keyVarLogN               
 $ keyVarObs                
 $ stockRecruitmentModelCode
 $ noScaledYears            
 $ keyScaledYears           
 $ keyParScaledYA           
 $ fbarRange                
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(canmackConf)
## maybe str(canmackConf) ; plot(canmackConf) ...
}
\keyword{datasets}
