\name{canmackData}
\alias{canmackData}
\docType{data}
\title{
canmackData
}
\description{
canmackData
}
\usage{data("canmackData")}
\format{
  The format is:
 $ noFleets    
 $ fleetTypes  
 $ sampleTimes 
 $ noYears     
 $ years       
 $ nobs        
 $ idx1        
 $ idx2        
 $ aux
 $ logobs         
 $ propMat   
 $ stockMeanWeight          
 $ catchMeanWeight          
 $ natMor                   
 $ landFrac                 
 $ disMeanWeight            
 $ landMeanWeight           
 $ propF                    
 $ propM                    
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(canmackData)
## maybe str(canmackData) ; plot(canmackData) ...
}
\keyword{datasets}
