\name{canmackParameters}
\alias{canmackParameters}
\docType{data}
\title{
canmackParameters
}
\description{
canmackParameters
}
\usage{data("canmackParameters")}
\format{
  The format is:
List of 14
 $ logFpar     
 $ logQpow     
 $ logSdLogFsta
 $ logSdLogN   
 $ logSdLogObs 
 $ rec_loga    
 $ rec_logb    
 $ itrans_rho   
 $ logScale    
 $ logScaleSSB 
 $ logPowSSB   
 $ logSdSSB    
 $ logF        
 $ logN        
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(canmackParameters)
## maybe str(canmackParameters) ; plot(canmackParameters) ...
}
\keyword{datasets}
