% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{catchplot}
\alias{catchplot}
\alias{catchplot.default}
\title{CCAM catch plot}
\usage{
catchplot(x, fleet = NULL, ...)

\method{catchplot}{default}(x, fleet = NULL, ...)
}
\arguments{
\item{x}{the object(s) returned from ccam.fit or forecast}

\item{fleet}{add observations from this fleet}

\item{...}{extra arguments transferred to plotit}
}
\description{
CCAM catch plot
}
\details{
Plot of spawning stock biomass
}
