% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{catchtable}
\alias{catchtable}
\alias{catchtable.ccam}
\alias{catchtable.ccamset}
\title{Catch table}
\usage{
catchtable(fit, obs.show = FALSE, ...)

\method{catchtable}{ccam}(fit, obs.show = FALSE, ...)

\method{catchtable}{ccamset}(fit, obs.show = FALSE, ...)
}
\arguments{
\item{fit}{...}

\item{obs.show}{logical add a column with catch sum of product rowsums(C*W)}

\item{...}{extra arguments not currently used}
}
\description{
Catch table
}
\details{
...
}
