% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.ccam}
\alias{coef.ccam}
\title{Extract fixed coefficients of ccam object}
\usage{
\method{coef}{ccam}(object, ...)
}
\arguments{
\item{object}{ccam fitted object (result from ccam.fit)}

\item{...}{extra arguments}
}
\description{
Extract fixed coefficients of ccam object
}
\details{
...
}
