% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp.R
\name{copy}
\alias{copy}
\title{copy a list (MPs or OMs)}
\usage{
copy(x, n, name)
}
\arguments{
\item{x}{list to copy}

\item{n}{number of times to copy. If vector than the number of copies equals the sum of n (with the i'th element corresponding to name i)}

\item{name}{name basis for the list copies}
}
\description{
copy a list (MPs or OMs)
}
\details{
list names are name plus a number from 1 to n, OM/MP label is automatically adapted
}
