% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{corplot}
\alias{corplot}
\alias{corplot.ccam}
\alias{corplot.ccamres}
\title{Plots between-age correlations by fleet, either estimated or empirical using residuals.}
\usage{
corplot(x, ...)

\method{corplot}{ccam}(x, ...)

\method{corplot}{ccamres}(x, ...)
}
\arguments{
\item{x}{Either a ccam fit as returned by s=ccam.fit OR the object returned from residuals.ccam}

\item{...}{extra arguments to plot}
}
\description{
Plots between-age correlations by fleet, either estimated or empirical using residuals.
}
