% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{crl}
\alias{crl}
\title{continuous ratio logit transformation of catch-at-age matrix}
\usage{
crl(x)
}
\arguments{
\item{x}{catch-at-age vector, dataframe or matrix (rows=ages, columns=years).}
}
\description{
continuous ratio logit transformation of catch-at-age matrix
}
\details{
(c) NOEL CADIGAN, adapted by EVB for use on vectors
}
