% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{crlInverse}
\alias{crlInverse}
\title{continuous ratio logit transformed matrix of catches-at-age to matrix of proportions}
\usage{
crlInverse(x)
}
\arguments{
\item{x}{catch-at-age matrix (crl transformed, rows=ages, columns=years).}
}
\description{
continuous ratio logit transformed matrix of catches-at-age to matrix of proportions
}
\details{
...
}
