% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{cumsum.bounded}
\alias{cumsum.bounded}
\title{cumsum.bounded}
\usage{
\method{cumsum}{bounded}(x, lower = 0, upper = 500)
}
\arguments{
\item{x}{vector of values to cumsum}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\description{
cumsum.bounded
}
\details{
cumsum function but with bounds. Used to create a bounded random walk.
}
