% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{defcon}
\alias{defcon}
\title{Setup basic minimal configuration for ccam assessment}
\usage{
defcon(dat)
}
\arguments{
\item{dat}{ccam data object}
}
\value{
a list containing
\item{minAge}{}
\item{maxAge}{}
\item{maxAgePlusGroup}{}
\item{keyLogFsta}{}
\item{keyLogFpar}{}
\item{keyQpow}{}
\item{keyVarF}{}
\item{keyVarLogN}{}
\item{keyVarObs}{}
\item{obsCorStruct}{}
\item{keyCorObs}{}
\item{corFlag}{set to zero as a placeholder here.}
\item{stockRecruitmentModelCode}{set to zero as a placeholder here.}
\item{noScaledYears}{set to zero as a placeholder here.}
\item{keyScaledYears}{a scalar set to zero as a placeholder here.}
\item{keyParScaledYA}{an array set to zero as a placeholder here.}
\item{fbarRange}{}
\item{obsLikelihoodFlag}{A vector with an element for each fleet. "LN","ALN" or for fleetType 3 "CE" or "RO"}
}
\description{
Setup basic minimal configuration for ccam assessment
}
\details{
...
}
