% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{diamondplot}
\alias{diamondplot}
\alias{diamondplot.ccamforecast}
\alias{diamondplot.forecastset}
\title{Plots fit to data}
\usage{
diamondplot(x, what, ylab = what, xlab = "Management Procedure",
  hline = NULL, year)

\method{diamondplot}{ccamforecast}(x, ...)

\method{diamondplot}{forecastset}(x, what, ylab = what,
  xlab = "Management Procedure", hline = NULL, year = NULL)
}
\arguments{
\item{x}{the object(s) returned from ccam.forecast}

\item{what}{statistic returned from ccam.forecast}

\item{ylab}{y label}

\item{xlab}{x label}

\item{hline}{plot horizontal line(s)}

\item{year}{year for which the statistics needs to be plotted}
}
\description{
Plots fit to data
}
\details{
performance (what) by management procedure, IE and OM type. OM type is defined by the OM name (numbers and 'OM' are removed).
}
