% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{expplot}
\alias{expplot}
\alias{expplot.default}
\title{CCAM exploitation rate plot}
\usage{
expplot(x, ...)

\method{expplot}{default}(x, ...)
}
\arguments{
\item{x}{the object(s) returned from ccam.fit}

\item{...}{extra arguments transferred to plotit}
}
\description{
CCAM exploitation rate plot
}
\details{
Plot of total stock biomass
}
