% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{extract}
\alias{extract}
\alias{extract.ccamforecast}
\alias{extract.forecastset}
\title{extract predictions for a given variable}
\usage{
extract(x, ...)

\method{extract}{ccamforecast}(x, what)

\method{extract}{forecastset}(x, what)
}
\arguments{
\item{x}{A forecast object}

\item{...}{extra arguments}
}
\description{
extract predictions for a given variable
}
\details{
...
}
