% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{faytable}
\alias{faytable}
\alias{faytable.ccam}
\alias{faytable.ccamset}
\title{F-at-age table}
\usage{
faytable(fit, ...)

\method{faytable}{ccam}(fit, ...)

\method{faytable}{ccam}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
F-at-age table
}
\details{
...
}
