% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{fleet3}
\alias{fleet3}
\title{fleet 3}
\usage{
fleet3(x, fit, f, simpara, deterministic, ULpool = NULL, nm = nm, sw = sw,
  mo = mo, pm = pm, pf = pf, cw = cw)
}
\arguments{
\item{x}{simulations of states}

\item{fit}{object from ccam.fit}

\item{f}{fleet (numeric)}

\item{deterministic}{logistic}
}
\value{
returns an array of dim (na-1) x 2 (observation columns) x number of simulations
}
\description{
fleet 3
}
\details{
generates observations of fleet 3 (keybiomasstreat 4 not yet implemented, 3 needs to be verified)
}
