% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{fleet6}
\alias{fleet6}
\title{fleet 6}
\usage{
fleet6(x, nm)
}
\arguments{
\item{x}{simulations of states}

\item{nm}{simulation of natural mortality}
}
\value{
returns an array of dim (na-1) x 2 (observation columns) x number of simulations
}
\description{
fleet 6
}
\details{
generates observations of fleet 6 (deterministic)
}
