% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{force.fit}
\alias{force.fit}
\title{Force fit}
\usage{
force.fit(data, conf, parameters, nojit = 10, silent = TRUE, ...)
}
\arguments{
\item{data}{data to fit a model}

\item{conf}{model configuration}

\item{parameters}{initial parameter values}

\item{nojit}{maximum number of tries with different initial values}

\item{silent}{logical}

\item{...}{agruments to ccam.fit}
}
\value{
ccam object
}
\description{
Force fit
}
\details{
Try fitting the model with different initial parameters untill it fits (or untill it is clear it most likely never will)
}
