% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{forecast function to do shortterm}
\usage{
forecast(fit, fscale = NULL, catchval = NULL, fval = NULL, MP = NULL,
  nosim = 1000, year.base = max(fit$data$years),
  ave.years = max(fit$data$years) + (-4:0), rec.years = max(fit$data$years)
  + (-9:0), rec.meth = 5, rec.scale = 1, MPlabel = NULL, OMlabel = NULL,
  overwriteSelYears = NULL, deterministic = FALSE, CZ = 0, HZ = 0,
  IE = NULL, capLower = NULL, capUpper = NULL,
  UL.years = max(fit$data$years) + (-4:0), TAC.base = 0, bio.scale = NULL,
  verbose = TRUE)
}
\arguments{
\item{fit}{an assessment object of type ccam, as returned from the function ccam.fit}

\item{fscale}{a vector of f-scales. See details.}

\item{catchval}{a vector of target catches. See details.}

\item{fval}{a vector of target f values. See details.}

\item{MP}{a vector of management strategie names. see avail('MP')}

\item{nosim}{number of simulations default is 1000}

\item{year.base}{starting year default last year in assessment. Currently it is only supported to use last assessment year or the year before}

\item{ave.years}{vector of years to average for weights, maturity, M and such}

\item{rec.years}{vector of years to use to resample recruitment from}

\item{rec.meth}{methode to project recruitment}

\item{rec.scale}{scale recruitment (default is 1)}

\item{overwriteSelYears}{if a vector of years is specified, then the average selectivity of those years is used (not recommended)}

\item{deterministic}{option to turn all process noise off (not recommended, as it will likely cause bias)}

\item{CZ}{critical zone (to calculate probability ssb above)}

\item{HZ}{Healty zone (to calculate probability ssb above)}

\item{IE}{implementation error (vector). See avail('IE')}

\item{capLower}{lower limit of TAC (before IE)}

\item{capUpper}{upper limit of TAC (before IE)}

\item{UL.years}{vector of years to use to resample upper/lower catch bound ranges from}

\item{TAC.base}{TAC last year}

\item{bio.scale}{names list with multipliers used to scale future bio values (nm, en, pr,sw,cw,lf,dw,lw,pm,pf)}

\item{label}{optional label to appear in short table}
}
\value{
an object of type ccamforecast
}
\description{
forecast function to do shortterm
}
\details{
There are four ways to specify a scenario; fscale, catchval, fval and MP.
}
