% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{foreplot}
\alias{foreplot}
\alias{foreplot.ccamforecast}
\alias{foreplot.forecastset}
\title{Plots fit to data}
\usage{
foreplot(fit, what, ylab = "value", xlab = "Year", zone = NULL,
  ci = TRUE, ...)

\method{foreplot}{ccamforecast}(...)

\method{foreplot}{forecastset}(...)
}
\arguments{
\item{fit}{the object returned from ccam.forecast}

\item{what}{statistic returned from ccam.forecast}

\item{ylab}{y label}

\item{xlab}{x label}

\item{zone}{color zone above this y value}

\item{ci}{logical (ci are quantile ranges)}

\item{...}{extra arguments to plot}
}
\description{
Plots fit to data
}
\details{
Plot of probability of what over time
}
