% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp.R
\name{gmean}
\alias{gmean}
\title{geometric mean}
\usage{
gmean(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\description{
geometric mean
}
\details{
Geometric mean of a vector. Indicates the central tendency or typical value of a set of numbers by using the product of their values (as opposed to the arithmetic mean which uses their sum). The geometric mean is defined as the nth root of the product of n numbers.
}
