% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{heat}
\alias{heat}
\title{Heat plot}
\usage{
heat(x, high = NULL, low = NULL, xlab = "Year", ylab = "Age",
  posneg = FALSE)
}
\arguments{
\item{x}{matrix}

\item{high}{colors that indicate the higher spectrum}

\item{low}{colors that indicate the lower spectrum}

\item{ncol}{number of colors to use in plot}

\item{leg}{logical (plot legend?)}

\item{leground}{number of digits to round legend}
}
\description{
Heat plot
}
