% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEconstant}
\alias{IEconstant}
\title{IEconstant}
\usage{
IEconstant(x, y, value = 0)
}
\arguments{
\item{x}{number of simulations}

\item{y}{number of timestep}

\item{value}{value of IE (same unit as catch)}
}
\description{
IEconstant
}
\details{
keeps TAC as it is
}
