% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEdep2550}
\alias{IEdep2550}
\title{IEdep2550}
\usage{
IEdep2550(x, y, seed = NULL, ssb, i)
}
\arguments{
\item{x}{number of simulations}

\item{y}{number of timestep}

\item{seed}{set.seed}

\item{ssb}{vector of ssb values}

\item{i}{future year index}
}
\description{
IEdep2550
}
\details{
generates matrix of size nosim x i
}
