% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEdep75100}
\alias{IEdep75100}
\title{IEdep75100}
\usage{
IEdep75100(x, y, seed = NULL, ssb, TAC, i, past = NULL)
}
\arguments{
\item{x}{number of simulations}

\item{y}{number of timestep}

\item{seed}{set.seed}

\item{ssb}{vector of ssb values}

\item{TAC}{TAC values}

\item{i}{future year index}
}
\description{
IEdep75100
}
\details{
generates matrix of size nosim x i
}
