% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEdepcopy}
\alias{IEdepcopy}
\title{IEdepcopy}
\usage{
IEdepcopy(x, y, seed = NULL, ssb, TAC, i, past = NULL)
}
\arguments{
\item{x}{number of simulations}

\item{y}{number of timestep}

\item{seed}{set.seed}

\item{ssb}{vector of ssb values}

\item{TAC}{TAC values}

\item{i}{future year index}
}
\description{
IEdepcopy
}
\details{
generates matrix of size nosim x i. US copies Canadian TAC and catches 25-50% of northern contingent.
}
