% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEgamma1}
\alias{IEgamma1}
\title{IEgamma1}
\usage{
IEgamma1(TAC, i)
}
\arguments{
\item{TAC}{TAC based on MP}

\item{i}{numeric between 0 and time series length-1 indicating the timestep}
}
\description{
IEgamma1
}
\details{
adds catch to TAC based on normal distribution
}
