% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEindep2019}
\alias{IEindep2019}
\title{IEindep2019}
\usage{
IEindep2019(x, y, seed = NULL)
}
\arguments{
\item{x}{number of simulations}

\item{y}{number of timestep}

\item{seed}{set.seed}
}
\description{
IEindep2019
}
\details{
generates matrix nosim x nyears with implementation errors that will be added to the TAC00
}
