% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEindep500}
\alias{IEindep500}
\title{IEindep500}
\usage{
IEindep500(x, y, seed = NULL)
}
\arguments{
\item{x}{number of simulations}

\item{y}{number of timestep}

\item{seed}{set.seed}
}
\description{
IEindep500
}
\details{
generates matrix nosim x nyears with implementation errors that will be added to the TAC00
}
