% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEindep6000}
\alias{IEindep6000}
\title{IEindep6000}
\usage{
IEindep6000(x, y, seed = NULL)
}
\arguments{
\item{x}{number of simulations}

\item{y}{number of timestep}

\item{seed}{set.seed}
}
\description{
IEindep6000
}
\details{
generates matrix nosim x nyears with implementation errors that will be added to the TAC
}
