% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEindep7200}
\alias{IEindep7200}
\title{IEindep7200}
\usage{
IEindep7200(x, y, seed = NULL)
}
\arguments{
\item{x}{number of simulations}

\item{y}{number of timestep}

\item{seed}{set.seed}
}
\description{
IEindep7200
}
\details{
generates matrix nosim x nyears with implementation errors that will be added to the TAC
}
