% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEindepdecr}
\alias{IEindepdecr}
\title{IEindepdecr}
\usage{
IEindepdecr(x, y, seed = NULL)
}
\arguments{
\item{x}{number of simulations}

\item{y}{number of timestep}

\item{seed}{set.seed}
}
\description{
IEindepdecr
}
\details{
generates matrix nosim x nyears with implementation errors that will be added to the TAC00
}
