% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEindepdecrsteep}
\alias{IEindepdecrsteep}
\title{IEindepdecrsteep}
\usage{
IEindepdecrsteep(x, y, seed = NULL)
}
\arguments{
\item{x}{number of simulations}

\item{y}{number of timestep}

\item{seed}{set.seed}
}
\description{
IEindepdecrsteep
}
\details{
generates matrix nosim x nyears with implementation errors that will be added to the TAC
}
