% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEnorm1}
\alias{IEnorm1}
\title{IEnorm1}
\usage{
IEnorm1(TAC, i)
}
\arguments{
\item{TAC}{TAC based on MP}

\item{i}{numeric between 0 and time series length-1 indicating the timestep}
}
\description{
IEnorm1
}
\details{
adds catch to TAC based on normal distribution: constant 6000 (sd 2000)
}
