% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEnorm2}
\alias{IEnorm2}
\title{IEnorm2}
\usage{
IEnorm2(TAC, i)
}
\arguments{
\item{TAC}{TAC based on MP}

\item{i}{numeric between 0 and time series length-1 indicating the timestep}
}
\description{
IEnorm2
}
\details{
adds catch to TAC based on normal distribution
}
