% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEnorm3}
\alias{IEnorm3}
\title{IEnorm3}
\usage{
IEnorm3(TAC, i)
}
\arguments{
\item{TAC}{TAC based on MP}

\item{i}{numeric between 0 and time series length-1 indicating the timestep}
}
\description{
IEnorm3
}
\details{
adds catch to TAC based on normal distribution
}
