% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEnormgamma1}
\alias{IEnormgamma1}
\title{IEnormgamma1}
\usage{
IEnormgamma1(TAC, i)
}
\arguments{
\item{TAC}{TAC based on MP}

\item{i}{numeric between 0 and time series length-1 indicating the timestep}
}
\description{
IEnormgamma1
}
\details{
adds catch to TAC based on normal distribution: constant 6000 (sd 2000) + gamma
}
