% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEnormgamma2}
\alias{IEnormgamma2}
\title{IEnormgamma2}
\usage{
IEnormgamma2(TAC, i)
}
\arguments{
\item{TAC}{TAC based on MP}

\item{i}{numeric between 0 and time series length-1 indicating the timestep}
}
\description{
IEnormgamma2
}
\details{
adds catch to TAC based on normal distribution: constant 7200 (sd 2400) + gamma
}
