% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEnormgamma3}
\alias{IEnormgamma3}
\title{IEnormgamma3}
\usage{
IEnormgamma3(TAC, i)
}
\arguments{
\item{TAC}{TAC based on MP}

\item{i}{numeric between 0 and time series length-1 indicating the timestep}
}
\description{
IEnormgamma3
}
\details{
adds catch to TAC based on normal distribution: gradualy reduction first 3 years to 3000 + gamma
}
