% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEnormgamma4}
\alias{IEnormgamma4}
\alias{IEnormgamma5}
\title{IEnormgamma4}
\usage{
IEnormgamma4(TAC, i)

IEnormgamma5(TAC, i)
}
\arguments{
\item{TAC}{TAC based on MP}

\item{i}{numeric between 0 and time series length-1 indicating the timestep}

\item{TAC}{TAC based on MP}

\item{i}{numeric between 0 and time series length-1 indicating the timestep}
}
\description{
IEnormgamma4

IEnormgamma5
}
\details{
adds catch to TAC based on normal distribution: gradual reductino first 6 years to 1000 + gamma

adds catch to TAC based on normal distribution: constant 4800 (20% decrease) + gamma
}
