% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEnothing}
\alias{IEnothing}
\title{IEnothing}
\usage{
IEnothing(x, y, seed = NULL)
}
\arguments{
\item{x}{number of simulations}

\item{y}{number of timestep}

\item{seed}{set.seed}
}
\description{
IEnothing
}
\details{
keeps TAC as it is
}
