% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{IEplot}
\alias{IEplot}
\title{Plot implementation error}
\usage{
IEplot(IEmeans, IEsds, col = c("#000000", "#88CCEE", "#44AA99", "#117733",
  "#999933", "#DDCC77", "#661100", "#CC6677", "#882255", "#AA4499"))
}
\arguments{
\item{IEmeans}{list of mean values}

\item{IEsds}{list of sd values}

\item{col}{colors}
}
\description{
Plot implementation error
}
