% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{invcrl}
\alias{invcrl}
\title{continuous ratio logit transformed matrix of catches-at-age to matrix of proportions}
\usage{
invcrl(x)
}
\arguments{
\item{x}{catch-at-age  (if matrix crl transformed, rows=ages, columns=years).}
}
\description{
continuous ratio logit transformed matrix of catches-at-age to matrix of proportions
}
\details{
...
}
