% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{jit}
\alias{jit}
\title{Jitter runs}
\usage{
jit(data, conf, par = defpar(data, conf), nojit = 10, sd = 0.25,
  ncores = detectCores(), parallell = TRUE, ...)
}
\arguments{
\item{data}{data to fit a model}

\item{conf}{model configuration}

\item{par}{initial values to jitter around. The defaule ones are returned from the defpar function}

\item{nojit}{a list of vectors. Each element in the list specifies a run where the fleets mentioned are omitted}

\item{sd}{the standard deviation used to jitter the initial values (most parameters are on a log scale, so similar to cv)}

\item{ncores}{the number of cores to attemp to use}
}
\value{
A "ccamset" object, which is basically a list of ccam fits
}
\description{
Jitter runs
}
\details{
...
}
