% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{kobeplot}
\alias{kobeplot}
\alias{kobeplot.ccam}
\title{Kobe plot}
\usage{
kobeplot(x)

\method{kobeplot}{ccam}(x, textsize = 2, limit = NULL, legend = FALSE)
}
\arguments{
\item{x}{fit from ccam}
}
\description{
Kobe plot
}
\details{
Instead of msy F40% is used as a proxy
}
