% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.ccam}
\alias{logLik.ccam}
\title{Log likelihood of ccam object}
\usage{
\method{logLik}{ccam}(object, ...)
}
\arguments{
\item{object}{ccam fitted object (result from ccam.fit)}

\item{...}{extra arguments}
}
\description{
Log likelihood of ccam object
}
\details{
...
}
