% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{matplot2}
\alias{matplot2}
\title{matplot2 plot}
\usage{
matplot2(x, ylab = "Value", xlab = "Year", col = NULL, lwd = 1)
}
\arguments{
\item{x}{matrix}

\item{ylab}{y label}

\item{xlab}{x label}
}
\description{
matplot2 plot
}
