% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{modelplot}
\alias{modelplot}
\alias{modelplot.default}
\title{CCAM model plot (ssb, F, rec)}
\usage{
modelplot(x)

\method{modelplot}{default}(x)
}
\arguments{
\item{x}{the object(s) returned from ccam.fit or forecast}
}
\description{
CCAM model plot (ssb, F, rec)
}
\details{
Plot of spawning stock biomass
}
