% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{modeltable}
\alias{modeltable}
\alias{modeltable.ccam}
\alias{modeltable.ccamset}
\alias{modeltable.forecastset}
\title{model table}
\usage{
modeltable(fits, ...)

\method{modeltable}{ccam}(fits, ...)

\method{modeltable}{ccamset}(fits, ...)

\method{modeltable}{forecastset}(fits, ...)
}
\arguments{
\item{fits}{A ccam fit as returned from the ccam.fit function, or a collection c(fit1, fit2, ...) of such fits}

\item{...}{extra arguments not currently used}
}
\description{
model table
}
\details{
...
}
