% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{MPccam}
\alias{MPccam}
\title{MPccam}
\usage{
MPccam(fit, data, parameters, TAC.base, i)
}
\arguments{
\item{fit}{output from ccam.fit}

\item{data}{list with data objects}

\item{parameters}{list of parameters objects}

\item{TAC.base}{TAC of previous year}

\item{i}{year in the future}
}
\description{
MPccam
}
\details{
Determine TAC by fiting censored model, determining F40% and health zones and applying F based on the latter. Attributed alternative data  can be used in case simulated data does not lead to convergence.
}
