% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{MPeggcomplex}
\alias{MPeggcomplex}
\title{MPeggcomplex}
\usage{
MPeggcomplex(data, TAC.base)
}
\arguments{
\item{data}{list with data objects}

\item{TAC.base}{TAC of previous year}

\item{i}{year in the future}
}
\description{
MPeggcomplex
}
\details{
Based on the Gerromont and Butterworth rule.
}
