% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{MPeggcomplex0}
\alias{MPeggcomplex0}
\title{MPeggcomplex0}
\usage{
MPeggcomplex0(data, TAC.base)
}
\arguments{
\item{data}{list with data objects}

\item{TAC.base}{TAC of previous year}

\item{i}{year in the future}
}
\description{
MPeggcomplex0
}
\details{
Based on the Gerromont and Butterworth rule.
}
