% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{MPeggcomplex15000}
\alias{MPeggcomplex15000}
\title{MPeggcomplex15000}
\usage{
MPeggcomplex15000(data, TAC.base)
}
\arguments{
\item{data}{list with data objects}

\item{TAC.base}{TAC of previous year}

\item{i}{year in the future}
}
\description{
MPeggcomplex15000
}
\details{
Based on the Gerromont and Butterworth rule.
}
