% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{MPeggcomplex8000}
\alias{MPeggcomplex8000}
\title{MPeggcomplex8000}
\usage{
MPeggcomplex8000(data, TAC.base)
}
\arguments{
\item{data}{list with data objects}

\item{TAC.base}{TAC of previous year}

\item{i}{year in the future}
}
\description{
MPeggcomplex8000
}
\details{
Based on the Gerromont and Butterworth rule.
}
