% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{MPeggcomplexinterim}
\alias{MPeggcomplexinterim}
\title{MPeggcomplexinterim}
\usage{
MPeggcomplexinterim(data, TAC.base)
}
\arguments{
\item{data}{list with data objects}

\item{TAC.base}{TAC of previous year}

\item{i}{year in the future}
}
\description{
MPeggcomplexinterim
}
\details{
Based on the Gerromont and Butterworth rule.
}
