% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{MPeggsurvey}
\alias{MPeggsurvey}
\title{MPeggsurvey}
\usage{
MPeggsurvey(data, TAC.base)
}
\arguments{
\item{data}{list with data objects}

\item{TAC.base}{TAC of previous year}
}
\description{
MPeggsurvey
}
\details{
simply changes TAC in function of change in I (last year/year before). Change can not be bigger smaller than 0.5x or 2x.
}
