% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{MPeggsurveytrail3interim}
\alias{MPeggsurveytrail3interim}
\title{MPeggsurveytrail3interim}
\usage{
MPeggsurveytrail3interim(data, TAC.base, i)
}
\arguments{
\item{data}{list with data objects}

\item{TAC.base}{TAC of previous year}

\item{i}{year in the future}
}
\description{
MPeggsurveytrail3interim
}
\details{
simply changes TAC in function of change in I (last year/ geometric mean 3 years before). Change can not be bigger smaller than 0.5x or 2x.
}
