% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{MPspm}
\alias{MPspm}
\title{MPspm}
\usage{
MPspm(data, TAC.base, i)
}
\arguments{
\item{TAC.base}{TAC of previous state}

\item{fit}{output from ccam.fit}

\item{sim}{simulations of states}
}
\description{
MPspm
}
\details{
determine TAC by fiting surplus production model from SPiCT package
}
