% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{MSEplot}
\alias{MSEplot}
\alias{MSEplot.ccamforecast}
\alias{MSEplot.forecastset}
\title{Plot MSE criteria}
\usage{
MSEplot(x)

\method{MSEplot}{ccamforecast}(x, ...)

\method{MSEplot}{forecastset}(x)
}
\arguments{
\item{x}{the object(s) returned from forecast}
}
\description{
Plot MSE criteria
}
\details{
types are residuals ~ year (1); residuals ~ predicted (2); predicted ~ observed (3)

plots a table with the result for each objective (as TRUE/FALSE or a number) by OM, IE and MP. Objectives need to be adapted inside the plot code.
}
