% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{newobs}
\alias{newobs}
\title{create new observations (logobs), only for one new year}
\usage{
newobs(fit, sim, simpara, ULpool, deterministic, nm)
}
\arguments{
\item{fit}{returned from ccam.fit}

\item{sim}{simulation of states}

\item{ULpool}{matrix with ratios of predicted values vs the upper or lower bound}
}
\value{
returns and array of size (n observations,2 (lower and upper column), number of simulations)
}
\description{
create new observations (logobs), only for one new year
}
\details{
create new observations to apply model based management procedures
}
