% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs.ccam}
\alias{nobs.ccam}
\title{Extract number of observations from ccam object}
\usage{
\method{nobs}{ccam}(object, ...)
}
\arguments{
\item{object}{ccam fitted object (result from ccam.fit)}

\item{...}{extra arguments}
}
\description{
Extract number of observations from ccam object
}
\details{
...
}
