% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ntable}
\alias{ntable}
\alias{ntable.ccam}
\alias{ntable.ccamset}
\title{N table}
\usage{
ntable(fit, ...)

\method{ntable}{ccam}(fit, ...)

\method{ntable}{ccamset}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
N table
}
\details{
...
}
