% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{obscov}
\alias{obscov}
\alias{obscov.ccam}
\alias{obscov.ccamset}
\title{Extract observation covariance matrices from a CCAM fit}
\usage{
obscov(fit, corr = FALSE, ...)

\method{obscov}{ccam}(fit, corr = FALSE, ...)

\method{obscov}{ccamset}(fit, corr = FALSE, ...)
}
\arguments{
\item{fit}{the object returned from ccam.fit}

\item{corr}{if TRUE return correlation matrices rather than covariances}

\item{...}{extra arguments not currently used}
}
\value{
a list of matrices
}
\description{
Extract observation covariance matrices from a CCAM fit
}
