% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp.R
\name{.onLoad}
\alias{.onLoad}
\title{onload function}
\usage{
.onLoad(libname = find.package("CCAM"), pkgname = "CCAM")
}
\description{
onload function
}
\details{
stuff that need's to be done when loading the package
}
