% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{parplot}
\alias{parplot}
\alias{parplot.default}
\title{CCAM parameter plot}
\usage{
parplot(fit, cor.report.limit = 0.95, col = NULL)

\method{parplot}{default}(fit, cor.report.limit = 0.95, col = NULL)
}
\arguments{
\item{fit}{the object returned from ccam.fit}

\item{cor.report.limit}{correlations with absolute value > this number is reported in the plot}

\item{col}{colors}

\item{...}{extra arguments transferred to plot}
}
\description{
CCAM parameter plot
}
\details{
Plot of all estimated model parameters (fixed effects). Shown with confidence interval.
}
