% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{peplot}
\alias{peplot}
\alias{peplot.ccam}
\alias{peplot.ccamset}
\title{CCAM PE plot}
\usage{
peplot(x, heat = TRUE, ...)

\method{peplot}{ccam}(x, ...)

\method{peplot}{ccamset}(x, heat = TRUE, ...)
}
\arguments{
\item{x}{the object(s) returned from ccam.fit or forecast}

\item{heat}{logical. If true a heat plot is used, if false a bubble plot}

\item{...}{parameters used in heat() or bubble()}
}
\description{
CCAM PE plot
}
\details{
Plot of proces error
}
