% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ccam}
\alias{plot.ccam}
\title{Plot ccam object}
\usage{
\method{plot}{ccam}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments (not possible to use add=TRUE --- please collect to a list of fits using e.g the c(...), and then plot that collected object)}
}
\description{
Plot ccam object
}
\details{
...
}
