% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ccamforecast}
\alias{plot.ccamforecast}
\title{Plot ccamforecast object}
\usage{
\method{plot}{ccamforecast}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Plot ccamforecast object
}
\details{
...
}
