% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ccamset}
\alias{plot.ccamset}
\title{Plot ccam object}
\usage{
\method{plot}{ccamset}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Plot ccam object
}
\details{
...
}
