% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.forecastset}
\alias{plot.forecastset}
\title{Plot ccamforecast object}
\usage{
\method{plot}{forecastset}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Plot ccamforecast object
}
\details{
...
}
