% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotit}
\alias{plotit}
\alias{plotit.dfccam}
\alias{plotit.dfccamset}
\alias{plotit.dfccamforecast}
\alias{plotit.dfforecastset}
\title{Plot helper}
\usage{
plotit(x, ylab = "Estimate", xlab = "Year", ci = TRUE)

\method{plotit}{dfccam}(x, ylab = "Estimate", xlab = "Year", ci = TRUE,
  col = "black")

\method{plotit}{dfccamset}(x, ylab = "Estimate", xlab = "Year", ci = TRUE,
  col = c("#332288", "#88CCEE", "#44AA99", "#117733", "#999933", "#DDCC77",
  "#661100", "#CC6677", "#882255", "#AA4499"))

\method{plotit}{dfccamforecast}(x, ylab = "Estimate", xlab = "Year",
  ci = TRUE, col = "black")

\method{plotit}{dfforecastset}(x, ylab = "Estimate", xlab = "Year",
  ci = TRUE, col = c("#332288", "#88CCEE", "#44AA99", "#117733", "#999933",
  "#DDCC77", "#661100", "#CC6677", "#882255", "#AA4499"))
}
\arguments{
\item{x}{dataframe from table()}

\item{ylab}{y label}

\item{xlab}{x label}
}
\description{
Plot helper
}
\details{
The basic plotting used by many of the plotting functions (e.g. ssbplot, fbarplot ...)
}
