% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotobs}
\alias{plotobs}
\title{Observation plot}
\usage{
plotobs(x, trans = TRUE, years = unique(x$aux[, 1]),
  fleets = unique(x$aux[, 2]), type = "line", size = 2)
}
\arguments{
\item{x}{data object}

\item{trans}{logical. transform data?}

\item{years}{years to plot}

\item{fleets}{fleets to plot}

\item{type}{line, bar or area}
}
\description{
Observation plot
}
\details{
plot observations (logobs) by fleet
}
