% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prettymatplot}
\alias{prettymatplot}
\title{prettymatplot plot}
\usage{
prettymatplot(x, ylab = "Value", xlab = "Year", col = NULL, lwd = 1)
}
\arguments{
\item{x}{matrix}

\item{ylab}{y label}

\item{xlab}{x label}
}
\description{
prettymatplot plot
}
