% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ccamcoef}
\alias{print.ccamcoef}
\title{Print ccamcoef object}
\usage{
\method{print}{ccamcoef}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Print ccamcoef object
}
\details{
...
}
