% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ccamforecast}
\alias{print.ccamforecast}
\title{Print ccamforecast object}
\usage{
\method{print}{ccamforecast}(x, ...)
}
\arguments{
\item{x}{an object as returned from the forecast function}

\item{...}{extra arguments}
}
\description{
Print ccamforecast object
}
\details{
...
}
