% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ccamres}
\alias{print.ccamres}
\title{Print ccamres object}
\usage{
\method{print}{ccamres}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Print ccamres object
}
\details{
...
}
