% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ccamset}
\alias{print.ccamset}
\title{Print ccamset object}
\usage{
\method{print}{ccamset}(x, ...)
}
\arguments{
\item{x}{a list of ccam models}

\item{...}{extra arguments}
}
\description{
Print ccamset object
}
\details{
...
}
