% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ccamypr}
\alias{print.ccamypr}
\title{Print ccamypr object}
\usage{
\method{print}{ccamypr}(x, ...)
}
\arguments{
\item{x}{an object as returned from the ypr function}

\item{...}{extra arguments}
}
\description{
Print ccamypr object
}
\details{
...
}
