% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{procres}
\alias{procres}
\title{Compute process residuals (single joint sample)}
\usage{
procres(fit, ...)
}
\arguments{
\item{fit}{the fitted object as returned from the ccam.fit function}

\item{...}{extra arguments (not currently used)}
}
\value{
an object of class \code{ccamres}
}
\description{
Compute process residuals (single joint sample)
}
\details{
...
}
