% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prodplot}
\alias{prodplot}
\alias{prodplot.ccam}
\title{Plot surplus production vs SSB}
\usage{
prodplot(x)

\method{prodplot}{ccam}(x)
}
\arguments{
\item{x}{fit from ccam}
}
\description{
Plot surplus production vs SSB
}
