% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{raw}
\alias{raw}
\alias{raw.ccamforecast}
\alias{raw.forecastset}
\title{extract raw predictions for a given variable}
\usage{
raw(x, what)

\method{raw}{ccamforecast}(x, what)

\method{raw}{forecastset}(x, what)
}
\arguments{
\item{x}{A forecast object}

\item{what}{value from forecast}
}
\description{
extract raw predictions for a given variable
}
\details{
...
}
