% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{recplot}
\alias{recplot}
\alias{recplot.ccam}
\alias{recplot.ccamset}
\alias{recplot.ccamforecast}
\alias{recplot.forecastset}
\title{CCAM Recruits plot}
\usage{
recplot(fit, ...)

\method{recplot}{ccam}(fit, ...)

\method{recplot}{ccamset}(fit, ...)

\method{recplot}{ccamforecast}(fit, ...)

\method{recplot}{forecastset}(fit, ...)
}
\arguments{
\item{fit}{the object returned from ccam.fit}

\item{...}{extra arguments transferred to plot including the following: \cr
\code{add} logical, plotting is to be added on existing plot \cr
\code{ci} logical, confidence intervals should be plotted \cr
\code{cicol} color to plot the confidence polygon}
}
\description{
CCAM Recruits plot
}
\details{
Plot of numbers of recruits (youngest age class)
}
