% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rectable}
\alias{rectable}
\alias{rectable.default}
\title{Recruit table}
\usage{
rectable(fit, trans = exp, ...)

\method{rectable}{default}(fit, trans = exp, ...)
}
\arguments{
\item{fit}{...}

\item{trans}{exp by default}

\item{...}{extra arguments not currently used}
}
\description{
Recruit table
}
\details{
...
}
