% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.ccam}
\alias{residuals.ccam}
\title{Extract residuals from ccam object}
\usage{
\method{residuals}{ccam}(object, discrete = FALSE, ...)
}
\arguments{
\item{object}{ccam fitted object (result from ccam.fit)}

\item{discrete}{logical if model contain discrete observations}

\item{...}{extra arguments for TMB's oneStepPredict}
}
\description{
Extract residuals from ccam object
}
\details{
...
}
