% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{resplot}
\alias{resplot}
\alias{resplot.ccam}
\title{Plots residuals}
\usage{
resplot(fit, log = TRUE, ...)

\method{resplot}{ccam}(fit, trans = function(x) x,
  fleets = unique(fit$data$aux[, "fleet"]), type = 1, std = TRUE,
  out = 0, ...)
}
\arguments{
\item{fit}{the object returned from ccam.fit}

\item{...}{extra arguments to plot}

\item{fleets}{an integer vector of fleets to plot. Default is all of them}

\item{type}{Type of residual plot}

\item{std}{logical; standardize residuals?}

\item{out}{number of outliers to which to add identifier}
}
\description{
Plots residuals
}
\details{
types are residuals ~ year (1); residuals ~ predicted (2); predicted ~ observed (3)
}
