% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp.R
\name{savepng}
\alias{savepng}
\title{save pngs in one line}
\usage{
savepng(plot, wd, name, dim)
}
\arguments{
\item{plot}{plot}

\item{wd}{working directory in which to plase image}

\item{name}{file name}

\item{dim}{dimensions of png in cm}
}
\description{
save pngs in one line
}
