% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scplot}
\alias{scplot}
\alias{scplot.default}
\title{CCAM ssb and catch}
\usage{
scplot(x, ci = TRUE, ...)

\method{scplot}{default}(x, ci = TRUE, ...)
}
\arguments{
\item{x}{the object(s) returned from ccam.fit or forecast}

\item{ci}{logical (confidence interval)}

\item{...}{extra arguments transferred to plotit}
}
\description{
CCAM ssb and catch
}
\details{
Plot of spawning stock biomass
}
