% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{selplot}
\alias{selplot}
\alias{selplot.default}
\title{CCAM selectivity plot}
\usage{
selplot(x, ...)

\method{selplot}{default}(x, ...)
}
\arguments{
\item{x}{the object(s) returned from ccam.fit or forecast}

\item{...}{extra arguments transferred to plotit}
}
\description{
CCAM selectivity plot
}
\details{
Plot of selectivity
}
