% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{setup.ccam.data}
\alias{setup.ccam.data}
\title{Combine the data sources to CCAM readable object}
\usage{
setup.ccam.data(fleets = NULL, surveys = NULL, residual.fleet = NULL,
  total.catch = NULL, prop.mature = NULL, stock.mean.weight = NULL,
  catch.mean.weight = NULL, dis.mean.weight = NULL,
  land.mean.weight = NULL, natural.mortality = NULL, prop.f = NULL,
  prop.m = NULL, land.frac = NULL, recapture = NULL, env = NULL)
}
\arguments{
\item{fleets}{comm fleets vith effort (currently unimplemented)}

\item{surveys}{surveys}

\item{residual.fleet}{total catch minus commercial}

\item{prop.mature}{pm}

\item{stock.mean.weight}{sw}

\item{catch.mean.weight}{cw}

\item{dis.mean.weight}{dw}

\item{land.mean.weight}{lw}

\item{natural.mortality}{nm}

\item{prop.f}{...}

\item{prop.m}{...}

\item{land.frac}{...}

\item{recapture}{...}
}
\description{
Combine the data sources to CCAM readable object
}
\details{
...
}
