% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{simulate.ccam}
\alias{simulate.ccam}
\title{Simulate from a ccam object}
\usage{
\method{simulate}{ccam}(object, nsim = 1, seed = NULL,
  full.data = TRUE, ...)
}
\arguments{
\item{object}{ccam fitted object (result from ccam.fit)}

\item{nsim}{number of response lists to simulate. Defaults to 1.}

\item{seed}{random number seed}

\item{full.data}{logical, should each inner list contain a full list of data. Defaults to TRUE}

\item{...}{extra arguments}
}
\value{
returns a list of lists. The outer list has length \code{nsim}. Each inner list contains simulated values of \code{logF}, \code{logN}, and \code{obs} with dimensions equal to those parameters.
}
\description{
Simulate from a ccam object
}
\details{
...
}
