% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{spya}
\alias{spya}
\title{Standardized proportions at age}
\usage{
spya(tdat)
}
\arguments{
\item{x}{matrix to be standardized (years in columns, ages in rows)}
}
\description{
Standardized proportions at age
}
\details{
They show the age composition of the catch each year, and can be used to track cohorts trends. They show more clearly when catches are above or below average. Negative values are below average, positive values are above average.
}
