% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{srplot}
\alias{srplot}
\alias{srplot.ccam}
\title{Plots the stock recruitment}
\usage{
srplot(fit, ...)

\method{srplot}{ccam}(fit, textcol = "red", linecol = "black",
  add = FALSE, curve = FALSE, ...)
}
\arguments{
\item{fit}{the object returned from ccam.fit}

\item{...}{extra arguments to plot}

\item{textcol}{color of years on plot}

\item{add}{false if a new plot should be created}
}
\description{
Plots the stock recruitment
}
