% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ssb0plot}
\alias{ssb0plot}
\alias{ssb0plot.default}
\title{CCAM SSB0 plot}
\usage{
ssb0plot(x, ...)

\method{ssb0plot}{default}(x, ...)
}
\arguments{
\item{x}{the object(s) returned from ccam.fit or forecast}

\item{...}{extra arguments transferred to plotit}
}
\description{
CCAM SSB0 plot
}
\details{
Plot of spawning stock biomass
}
