% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ssb0table}
\alias{ssb0table}
\alias{ssb0table.default}
\title{SSB0 table}
\usage{
ssb0table(fit, trans = exp, ...)

\method{ssb0table}{default}(fit, trans = exp, ...)
}
\arguments{
\item{fit}{...}

\item{trans}{exp by default}

\item{...}{extra arguments not currently used}
}
\description{
SSB0 table
}
\details{
...
}
