% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ssbtable}
\alias{ssbtable}
\alias{ssbtable.default}
\title{SSB table}
\usage{
ssbtable(fit, trans = exp, ...)

\method{ssbtable}{default}(fit, trans = exp, ...)
}
\arguments{
\item{fit}{...}

\item{trans}{exp by default}

\item{...}{extra arguments not currently used}
}
\description{
SSB table
}
\details{
...
}
