% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{stepdat}
\alias{stepdat}
\title{add one year to data}
\usage{
stepdat(x, ob, aux, pr, sw, cw, nm, lf, dw, lw, pm, pf, en)
}
\arguments{
\item{x}{fit returned from ccam.fit}

\item{...}{next year data objects}
}
\description{
add one year to data
}
