% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp.R
\name{stick}
\alias{stick}
\title{stick function}
\usage{
stick(x, a, b, c, j0)
}
\arguments{
\item{x}{value of jy/yref}

\item{a}{parameter that scales the catch limit (hight of the curve)}

\item{b}{parameter that scales the catch limit (slope after j0)}

\item{c}{parameter determining the initial slope}

\item{j0}{parameter determining the bending point}
}
\description{
stick function
}
\details{
basis for an index based HCR (as in redfish)
}
