% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.ccam}
\alias{summary.ccam}
\title{Summary of ccam object}
\usage{
\method{summary}{ccam}(object, ...)
}
\arguments{
\item{object}{ccam fitted object (result from ccam.fit)}

\item{...}{extra arguments}
}
\description{
Summary of ccam object
}
\details{
...
}
