% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{surveyplot}
\alias{surveyplot}
\title{Plot surveys}
\usage{
surveyplot(x, col = c("#000000", "#88CCEE", "#44AA99", "#117733", "#999933",
  "#DDCC77", "#661100", "#CC6677", "#882255", "#AA4499"), xlim = NULL,
  ylim = NULL, ...)
}
\arguments{
\item{x}{surveys from read.icea}
}
\description{
Plot surveys
}
