% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tableset}
\alias{tableset}
\title{tableset}
\usage{
tableset(fit, fun, what, ...)
}
\arguments{
\item{fit}{list of returned objects from ccam.fit}

\item{fun}{table function to be applied to each list element}

\item{what}{variable}
}
\description{
tableset
}
