% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tradeplot}
\alias{tradeplot}
\alias{tradeplot.ccamforecast}
\alias{tradeplot.forecastset}
\title{trade plot for forecasts}
\usage{
tradeplot(fit, what.x, what.y, ...)

\method{tradeplot}{ccamforecast}(fit, what.x, what.y, add = FALSE,
  dotpch = 19, dotcex = 1.5, zone = NULL, col = "black", lab = TRUE,
  ci = FALSE, ylab = NULL, xlab = NULL, ylim = NULL, xlim = NULL, ...)

\method{tradeplot}{forecastset}(fit, what.x, what.y, leg = NULL,
  col = c("#332288", "#88CCEE", "#44AA99", "#117733", "#999933", "#DDCC77",
  "#661100", "#CC6677", "#882255", "#AA4499"), ylab = NULL, xlab = NULL,
  ylim = NULL, xlim = NULL, ...)
}
\arguments{
\item{fit}{the object returned from ccam.forecast}

\item{what.x}{what probability (ULR, Growth or Double)}

\item{what.y}{what variable (catch, catchcum, ...)}

\item{...}{extra arguments to plot}

\item{add}{logical, plotting is to be added on existing plot}

\item{dotpch}{numeric, pch}

\item{dotcex}{cex}

\item{zone}{color zone in plot green. numeric giving the boundary}

\item{col}{color of lines and dots}

\item{lab}{logical, add year labels?}

\item{ci}{add error bars (97.5/2.5 quantiles)}

\item{ylab}{y label}

\item{xlab}{x label}

\item{ylim}{y limits}

\item{xlim}{x limits}

\item{leg}{logical; plot legend}

\item{...}{additional parameters for plot}

\item{col}{color of lines and dots}

\item{ylab}{y label}

\item{xlab}{x label}

\item{ylim}{y limits}

\item{xlim}{x limits}

\item{...}{additional parameters for tradeplot.ccamforecast}
}
\description{
trade plot for forecasts
}
\details{
Plot of probability of what over time
}
