% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bubble}
\alias{bubble}
\alias{heat}
\title{Bubble plot}
\usage{
bubble(x, ylab = "Year", xlab = "Age", scale = 3, ...)

heat(x, col = c("darkgreen", "grey", "darkred"), xlab = "Year",
  ylab = "Age", ncol = 20, leg = TRUE, leground = 2, posneg = FALSE,
  ...)
}
\arguments{
\item{x}{matrix}

\item{scale}{scaler for the bubble size (cex size: sqrt(bubblevalue)*scale)}

\item{col}{colors that indicate the color scale limit}

\item{ncol}{number of colors to use in plot}

\item{leg}{logical (plot legend?)}

\item{leground}{number of digits to round legend}

\item{x}{matrix}
}
\description{
Bubble plot

Heat plot
}
