% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bubble}
\alias{bubble}
\title{Bubble plot}
\usage{
bubble(x, xlab = "Year", ylab = "Age", scale = 15, col = c("black",
  "darkgreen"), alpha = 0.8)
}
\arguments{
\item{x}{matrix}

\item{xlab}{x label}

\item{ylab}{y label}

\item{scale}{scaler for the bubble size (scale_range)}

\item{col}{color for the bubbles. if 2 colors are given, they indicate positive and negative values}

\item{alpha}{transparency of the bubbles}
}
\description{
Bubble plot
}
