% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{catchtable}
\alias{catchtable}
\alias{catchtable.default}
\title{Catch table}
\usage{
catchtable(fit, fleet = NULL, ...)

\method{catchtable}{default}(fit, fleet = NULL, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}

\item{obs.show}{logical add a column with catch sum of product rowsums(C*W)}
}
\description{
Catch table
}
\details{
...
}
