% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{defcon}
\alias{defcon}
\title{Setup basic minimal configuration for ccam assessment}
\usage{
defcon(dat)
}
\arguments{
\item{dat}{ccam data object}
}
\value{
a list containing
\item{minAge}{minimum age}
\item{maxAge}{maximum age}
\item{maxAgePlusGroup}{plus group (0 or 1)}
\item{keySel}{selectivity structure}
\item{keyLogFpar}{survey catchability}
\item{keyQpow}{Qpow structure}
\item{keyVarF}{F variance structure}
\item{keyVarLogN}{N variance structure}
\item{keyVarObs}{observation variance structure}
\item{obsCorStruct}{correlation structure if MVN observation error}
\item{keyCorObs}{correlation if MVN observation error}
\item{stockRecruitmentModelCode}{0: RW, 1: ricker, 2: BH, 3: mean}
\item{noScaledYears}{set to zero as a placeholder here.}
\item{keyScaledYears}{a scalar set to zero as a placeholder here.}
\item{keyParScaledYA}{an array set to zero as a placeholder here.}
\item{fbarRange}{Range of ages over which Fbar should be taken}
\item{obsLikelihoodFlag}{A vector with an element for each fleet. "LN","ALN" or for fleetType 3 "CE" or "RO"}
\item{fixVarToWeight}{a scalar set to zero as a placeholder here.}
}
\description{
Setup basic minimal configuration for ccam assessment
}
\details{
...
}
