% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{exptable}
\alias{exptable}
\alias{exptable.default}
\title{exploitation rate table}
\usage{
exptable(fit, ...)

\method{exptable}{default}(fit, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
exploitation rate table
}
\details{
...
}
