% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{extract}
\alias{extract}
\alias{extract.ccamforecast}
\alias{extract.forecastset}
\title{extract predictions for a given variable}
\usage{
extract(x, what, add = FALSE)

\method{extract}{ccamforecast}(x, what, add = FALSE)

\method{extract}{forecastset}(x, what, add = FALSE)
}
\arguments{
\item{x}{A forecast object}

\item{what}{value from forecast}

\item{...}{extra arguments}
}
\description{
extract predictions for a given variable
}
\details{
...
}
