% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fbarplot}
\alias{fbarplot}
\alias{fbarplot.default}
\title{CCAM fbar plot}
\usage{
fbarplot(x, ...)

\method{fbarplot}{default}(x, ...)
}
\arguments{
\item{x}{the object(s) returned from ccam.fit or forecast}

\item{...}{extra arguments transferred to plotit}
}
\description{
CCAM fbar plot
}
\details{
Plot of mean F
}
