% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fitplot}
\alias{fitplot}
\alias{fitplot.ccam}
\title{Plots fit to data}
\usage{
fitplot(fit, log = TRUE, ...)

\method{fitplot}{ccam}(fit, log = TRUE, fleets = unique(fit$data$aux[,
  "fleet"]), ...)
}
\arguments{
\item{fit}{the object returned from ccam.fit}

\item{log}{should the plot be against log-obs}

\item{...}{extra arguments to plot}

\item{fleets}{an integer vector of fleets to plot. Default is all of them}
}
\description{
Plots fit to data
}
