% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fitplot}
\alias{fitplot}
\alias{fitplot.default}
\title{fitplot}
\usage{
fitplot(x, type = "AIC")

\method{fitplot}{default}(x, type = c("AIC", "nll"))
}
\arguments{
\item{x}{fit from ccam}

\item{type}{'AIC' (default) or 'nll'}
}
\description{
fitplot
}
\details{
plots the AIC or nll for every model with the number of parameters on top
}
