% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{fleet6}
\alias{fleet6}
\title{fleet 6}
\usage{
fleet6(x, fit, f, simpara, deterministic, nm)
}
\arguments{
\item{x}{simulations of states}

\item{fit}{object from ccam.fit}

\item{f}{fleet (numeric)}

\item{deterministic}{logistic}
}
\value{
returns an array of dim (na-1)x2*nrow(sim)
}
\description{
fleet 6
}
\details{
generates observations of fleet 6
}
