% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{forecast function to do shortterm}
\usage{
forecast(fit, fscale = NULL, catchval = NULL, fval = NULL, MP = NULL,
  nosim = 1000, year.base = max(fit$data$years),
  ave.years = max(fit$data$years) + (-4:0), rec.years = max(fit$data$years)
  + (-9:0), rec.meth = 5, MPlabel = NULL, OMlabel = NULL,
  overwriteSelYears = NULL, deterministic = FALSE, CZ = 0, HZ = 0,
  IE = NULL, capLower = NULL, capUpper = NULL,
  UL.years = max(fit$data$years) + (-4:0), TAC.base = NULL)
}
\arguments{
\item{fit}{an assessment object of type ccam, as returned from the function ccam.fit}

\item{fscale}{a vector of f-scales. See details.}

\item{catchval}{a vector of target catches. See details.}

\item{fval}{a vector of target f values. See details.}

\item{nosim}{number of simulations default is 1000}

\item{year.base}{starting year default last year in assessment. Currently it is only supported to use last assessment year or the year before}

\item{ave.years}{vector of years to average for weights, maturity, M and such}

\item{rec.years}{vector of years to use to resample recruitment from}

\item{overwriteSelYears}{if a vector of years is specified, then the average selectivity of those years is used (not recommended)}

\item{deterministic}{option to turn all process noise off (not recommended, as it will likely cause bias)}

\item{CZ}{critical zone (to calculate probability ssb above)}

\item{HZ}{Healty zone (to calculate probability ssb above)}

\item{IE}{implementation error (list of IE per time step)}

\item{capLower}{lower limit of TAC (before IE)}

\item{capUpper}{upper limit of TAC (before IE)}

\item{TAC.base}{TAC last year}

\item{label}{optional label to appear in short table}
}
\value{
an object of type ccamforecast
}
\description{
forecast function to do shortterm
}
\details{
There are three ways to specify a scenario. If e.g. four F values are specified (e.g. fval=c(.1,.2,.3,4)), then the first value is used in the last assessment year (base.year), and the three following in the three following years. Alternatively F's can be specified by a scale, or a target catch. Only one option can be used per year. So for instance to set a catch in the first year and an F-scale in the following one would write catchval=c(10000,NA,NA,NA), fscale=c(NA,1,1,1). The length of the vector specifies how many years forward the scenarios run.
}
