% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{foreplot}
\alias{foreplot}
\alias{foreplot.ccamforecast}
\alias{foreplot.forecastset}
\title{Plots fit to data}
\usage{
foreplot(x, what.y, what.x = NULL, ylab = what.y, xlab = "Year",
  rect = NULL, ci = TRUE, vline = NULL, hline = NULL, by = NULL,
  text = FALSE)

\method{foreplot}{ccamforecast}(x, ...)

\method{foreplot}{forecastset}(x, what.y, what.x = NULL, ylab = what.y,
  xlab = "Year", rect = NULL, ci = TRUE, vline = NULL, hline = NULL,
  by = NULL, text = FALSE)
}
\arguments{
\item{x}{the object(s) returned from ccam.forecast}

\item{what.y}{statistic returned from ccam.forecast}

\item{what.x}{statistic returned from ccam.forecast}

\item{ylab}{y label}

\item{xlab}{x label}

\item{rect}{color zone above this y value}

\item{ci}{logical (ci are quantile ranges)}

\item{vline}{plot vertical line(s)}

\item{hline}{plot horizontal line(s)}

\item{by}{IE, MP, OM}

\item{text}{logical. plot year labels}

\item{...}{extra arguments to plot}
}
\description{
Plots fit to data
}
\details{
Plot of probability of what over time
}
