% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{foreplot}
\alias{foreplot}
\alias{foreplot.ccamforecast}
\alias{foreplot.forecastset}
\title{Plots forecasting results}
\usage{
foreplot(x, what.y, what.x = NULL, ylab = what.y, xlab = "Year",
  rect = NULL, ci = TRUE, vline = NULL, hline = NULL, by = NULL,
  text = FALSE, IE = NULL, year = NULL, legendnames = NULL,
  data = FALSE, ratio = FALSE)

\method{foreplot}{ccamforecast}(x, ...)

\method{foreplot}{forecastset}(x, what.y, what.x = NULL, ylab = what.y,
  xlab = "Year", rect = NULL, ci = TRUE, vline = NULL, hline = NULL,
  by = NULL, text = FALSE, IE = NULL, year = NULL, legendnames = NULL,
  data = FALSE, ratio = FALSE)
}
\arguments{
\item{x}{the object(s) returned from ccam.forecast}

\item{what.y}{statistic returned from ccam.forecast}

\item{what.x}{statistic returned from ccam.forecast}

\item{ylab}{y label}

\item{xlab}{x label}

\item{rect}{color zone above this y value}

\item{ci}{logical (ci are quantile ranges)}

\item{vline}{plot vertical line(s)}

\item{hline}{plot horizontal line(s)}

\item{by}{IE, MP, OM}

\item{text}{logical. plot year labels}

\item{IE}{numeric. which IE to use for naming if vector of IEs is used.}

\item{data}{logical data returned?}

\item{ratio}{logical. devide by LRP (0.4*SSBf40%)?}

\item{legend}{names vector of new legend names}
}
\description{
Plots forecasting results
}
\details{
Plot of probability of what over time
}
