% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{IEnothing}
\alias{IEnothing}
\title{IEnothing}
\usage{
IEnothing(TAC, i)
}
\arguments{
\item{TAC}{TAC based on MP}

\item{i}{numeric between 0 and time series length-1 indicating the timestep}
}
\description{
IEnothing
}
\details{
keeps TAC as it is
}
