% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{IEplot}
\alias{IEplot}
\title{Plot implementation error}
\usage{
IEplot(IEmeans, IEsds, col = NULL, ylab = "Undeclared catch (t)",
  linesize = 2)
}
\arguments{
\item{IEmeans}{list of mean values}

\item{IEsds}{list of sd values}

\item{col}{colors}

\item{ylab}{y label}

\item{linesize}{linesize}
}
\description{
Plot implementation error
}
