% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{MPccam}
\alias{MPccam}
\title{MPccam}
\usage{
MPccam(data, parameters, conf, TAC.base, i)
}
\arguments{
\item{data}{lists with data lists}

\item{parameters}{one list of parameters}

\item{conf}{one list with configurations}

\item{TAC.base}{TAC of previous year}

\item{i}{year in the future}
}
\description{
MPccam
}
\details{
Determine TAC by fiting censored model, determining F40% and health zones and applying F based on the latter. Attributed alternative data  can be used in case simulated data does not lead to convergence.
}
