% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ccamres}
\alias{plot.ccamres}
\title{Plot ccam residuals}
\usage{
\method{plot}{ccamres}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Plot ccam residuals
}
\details{
...
}
\examples{
\dontrun{
data(canmackData)
data(canmackConf)
data(canmackParameters)
fit <- ccam.fit(canmackData, canmackConf, canmackParameters)
par(ask=FALSE)
plot(residuals(fit))
}
}
