% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotit}
\alias{plotit}
\alias{plotit.dfccam}
\alias{plotit.dfccamset}
\alias{plotit.dfccamforecast}
\alias{plotit.dfforecastset}
\title{Plot helper}
\usage{
plotit(x, ylab = "Estimate", xlab = "Year", ci = TRUE,
  years = unique(x$year), linetype = 1, linesize = 2, ...)

\method{plotit}{dfccam}(x, ylab = "Estimate", xlab = "Year", ci = TRUE,
  years = unique(x$year), col = "black", linetype = 1, linesize = 2)

\method{plotit}{dfccamset}(x, ylab = "Estimate", xlab = "Year", ci = TRUE,
  years = unique(x$year), col = NULL, linetype = 1, linesize = 2,
  legendnames = NULL)

\method{plotit}{dfccamforecast}(x, ylab = "Estimate", xlab = "Year",
  ci = TRUE, years = unique(x$year), col = "black", linetype = 1,
  linesize = 2)

\method{plotit}{dfforecastset}(x, ylab = "Estimate", xlab = "Year",
  ci = TRUE, years = unique(x$year), col = NULL, linetype = 1,
  linesize = 2, legendnames = NULL)
}
\arguments{
\item{x}{dataframe from table()}

\item{ylab}{y label}

\item{xlab}{x label}

\item{ci}{logical (plot confidence interval?)}

\item{years}{years to plot}

\item{linetype}{linetype}

\item{linesize}{linesize}
}
\description{
Plot helper
}
\details{
The basic plotting used by many of the plotting functions (e.g. ssbplot, fbarplot ...)
}
