% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prettymatplot}
\alias{prettymatplot}
\title{prettymatplot plot}
\usage{
prettymatplot(x, ylab = ifelse(is.null(names(dimnames(x))[1]), "y",
  names(dimnames(x))[1]), xlab = ifelse(is.null(names(dimnames(x))[1]),
  "x", names(dimnames(x))[1]), legend = guide_legend(), col = NULL,
  lwd = 1)
}
\arguments{
\item{x}{matrix}

\item{ylab}{y label}

\item{xlab}{x label}

\item{legend}{logical}

\item{col}{colors}

\item{lwd}{line with}
}
\description{
prettymatplot plot
}
