% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{recplot}
\alias{recplot}
\alias{recplot.default}
\title{CCAM recruitment plot}
\usage{
recplot(x, trans = exp, ...)

\method{recplot}{default}(x, trans = exp, ...)
}
\arguments{
\item{x}{the object(s) returned from ccam.fit or forecast}

\item{trans}{exp by default}

\item{...}{extra arguments transferred to plotit}
}
\description{
CCAM recruitment plot
}
\details{
Plot of recruitment
}
