% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{seltable}
\alias{seltable}
\alias{seltable.default}
\title{Selectivity table}
\usage{
seltable(fit, fleet = NULL, ...)

\method{seltable}{default}(fit, fleet = NULL, ...)
}
\arguments{
\item{fit}{...}

\item{...}{extra arguments not currently used}
}
\description{
Selectivity table
}
\details{
...
}
