% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{spay}
\alias{spay}
\title{Standardized proportions at year}
\usage{
spay(tdat)
}
\arguments{
\item{x}{matrix to be standardized (years in columns, ages in rows)}
}
\description{
Standardized proportions at year
}
\details{
Standardized proportions at year can show cohort patterns more clearly.
}
