% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{srplot}
\alias{srplot}
\alias{srplot.ccam}
\alias{srplot.ccamset}
\title{Plots the stock recruitment}
\usage{
srplot(fit, textcol = "red", linecol = "black", curve = FALSE)

\method{srplot}{ccam}(fit, ...)

\method{srplot}{ccamset}(fit, textcol = "red", linecol = "black",
  curve = FALSE)
}
\arguments{
\item{fit}{the object returned from ccam.fit}

\item{textcol}{color of years on plot}

\item{linecol}{color of lines}

\item{curve}{add SR curve}
}
\description{
Plots the stock recruitment
}
